Move tool components
